moved ui check from rust to scripts. thank you stackoverflow
